Shortcode :: Articles List
- Move list/footer scripts into template files.

See: websharks/wp-kb-articles#5
